Remove clear shared URL cache when syncing
The cached images do not need to be cleared when syncing. This results in
unnecessary download of the same images each time a new LoL data dragon
version is released. If any new champions or champion skins are released, they
will be fetched and cached, but it is unnecessary to blow away the cache and
refetch everything.